requirements.txt
anthropic>=0.39.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
curl_cffi>=0.14.0
browser-cookie3>=0.20.0
